refactor: unify event handling and combine lib/bin
Switches to an event-driven update loop where the packets and
user input are handled via a single queue.  This is more efficient
than the polling approach originally used and simplifies the logic
a bit.

Moves the main TUI application out of the example directory and makes
it the primary `cargo run` target.  There isn't much use for the library
on its own as it stands today, so this makes it easier to run the
program.