feat: help screen, view toggles for period and dump
Adds a simple help screen with some of the keyboard commands which
are otherwise hidden.  This could stand to be better formatted using
a table with columns, etc.

Adds toggles for the period column in the message view, and for the
packet dump view.  These can be noisy/distracting when doing long
studies of playbacks, especially when comparing two traces, so are
good to be able to turn off.