trc: improve parsing, support multiple versions
Adds support for the 1.x and 2.x .trc file versions (except 1.2, which
seems to have only been released with PCAN-Explorer 5.0 Beta 1), and
refactors `TrcSource` to use this parser.  Includes tests against data
from the python-can repo example files.

Rolls the rust edition back to 2021 as we are not using any 2024
features just yet.